app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
awslambdaric==2.0.11
camelot-py==0.11.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
ghostscript==0.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.33.0
opencv-python==4.10.0.82
openpyxl==3.1.3
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypdf==4.2.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
simplejson==3.18.4
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
